+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 12.32.0
 
+* `[fmtutil]` Added method `PrettyDiff` for formatting diff numbers
 * `[fmtutil]` Fixed bug in `PrettyNum` with formatting negative numbers
 * `[fmtutil]` Code refactoring